Update shader precision
I still have no idea how much of a difference this makes, if any.
Apparently, it does on mobile, but most posts talking about it are from
10 years ago. And according to those same posts, iOS needs `highp` for
samplers, but, again, who knows.

+++ crates/flux/shaders/line.frag
@@ -6,10 +6,10 @@ in float vOpacity;
 
 layout(std140) uniform LineUniforms
 {
-  highp float uLineWidth;
-  highp float uLineLength;
-  highp float uLineBeginOffset;
-  highp float uLineFadeOutLength;
+  mediump float uLineWidth;
+  mediump float uLineLength;
+  mediump float uLineBeginOffset;
+  mediump float uLineFadeOutLength;
 };
 
 out vec4 fragColor;
